backend/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
gunicorn==20.1.0
orjson==3.8.4
aiokafka==0.8.0
PyJWT==2.6.0
motor==3.1.1
sentry-sdk==1.15.0
python-logstash==0.4.8
python-dotenv==0.21.0